copy graph node attributes to subgraph input node attributes
