+++ ewokscore/graph/__init__.py
@@ -1,0 +1,2 @@
+from .taskgraph import load_graph  # noqa F401
+from .taskgraph import TaskGraph  # noqa F401
