+++ ewokscore/tests/test_graph.py
@@ -1,5 +1,6 @@
 import pytest
 from ewokscore.graph import load_graph
+from ewokscore.graph.analysis import link_is_required
 
 
 def test_required_links():
@@ -14,46 +15,46 @@ def test_required_links():
         {"source": "source2a", "target": "source2b"},
         {"source": "source2b", "target": "target"},
     ]
-    graph = load_graph({"nodes": nodes, "links": links})
-    assert graph.link_is_required("source1", "target")
-    assert graph.link_is_required("source2a", "source2b")
-    assert graph.link_is_required("source2b", "target")
+    taskgraph = load_graph({"nodes": nodes, "links": links})
+    assert link_is_required(taskgraph.graph, "source1", "target")
+    assert link_is_required(taskgraph.graph, "source2a", "source2b")
+    assert link_is_required(taskgraph.graph, "source2b", "target")
 
     links[0]["conditions"] = [{"source_output": "a", "value": 1}]
-    graph = load_graph({"nodes": nodes, "links": links})
-    assert not graph.link_is_required("source1", "target")
-    assert graph.link_is_required("source2a", "source2b")
-    assert graph.link_is_required("source2b", "target")
+    taskgraph = load_graph({"nodes": nodes, "links": links})
+    assert not link_is_required(taskgraph.graph, "source1", "target")
+    assert link_is_required(taskgraph.graph, "source2a", "source2b")
+    assert link_is_required(taskgraph.graph, "source2b", "target")
     links[0].pop("conditions")
 
     links[1]["conditions"] = [{"source_output": "a", "value": 1}]
-    graph = load_graph({"nodes": nodes, "links": links})
-    assert graph.link_is_required("source1", "target")
-    assert not graph.link_is_required("source2a", "source2b")
-    assert not graph.link_is_required("source2b", "target")
+    taskgraph = load_graph({"nodes": nodes, "links": links})
+    assert link_is_required(taskgraph.graph, "source1", "target")
+    assert not link_is_required(taskgraph.graph, "source2a", "source2b")
+    assert not link_is_required(taskgraph.graph, "source2b", "target")
     links[1].pop("conditions")
 
     links[1]["conditions"] = [{"source_output": "a", "value": 1}]
     links[1]["required"] = True
-    graph = load_graph({"nodes": nodes, "links": links})
-    assert graph.link_is_required("source1", "target")
-    assert graph.link_is_required("source2a", "source2b")
-    assert graph.link_is_required("source2b", "target")
+    taskgraph = load_graph({"nodes": nodes, "links": links})
+    assert link_is_required(taskgraph.graph, "source1", "target")
+    assert link_is_required(taskgraph.graph, "source2a", "source2b")
+    assert link_is_required(taskgraph.graph, "source2b", "target")
     links[1].pop("conditions")
 
     links[2]["conditions"] = [{"source_output": "a", "value": 1}]
-    graph = load_graph({"nodes": nodes, "links": links})
-    assert graph.link_is_required("source1", "target")
-    assert graph.link_is_required("source2a", "source2b")
-    assert not graph.link_is_required("source2b", "target")
+    taskgraph = load_graph({"nodes": nodes, "links": links})
+    assert link_is_required(taskgraph.graph, "source1", "target")
+    assert link_is_required(taskgraph.graph, "source2a", "source2b")
+    assert not link_is_required(taskgraph.graph, "source2b", "target")
     links[2].pop("conditions")
 
     links[2]["conditions"] = [{"source_output": "a", "value": 1}]
     links[2]["required"] = True
-    graph = load_graph({"nodes": nodes, "links": links})
-    assert graph.link_is_required("source1", "target")
-    assert graph.link_is_required("source2a", "source2b")
-    assert graph.link_is_required("source2b", "target")
+    taskgraph = load_graph({"nodes": nodes, "links": links})
+    assert link_is_required(taskgraph.graph, "source1", "target")
+    assert link_is_required(taskgraph.graph, "source2a", "source2b")
+    assert link_is_required(taskgraph.graph, "source2b", "target")
     links[2].pop("conditions")
